requirements.txt
flask==3.1.0
pylint==3.3.2
python-dotenv==1.0.1
